main refactor
